Add mocks.go file
