+++ enterprise/internal/batches/service/service.go
@@ -638,13 +638,3 @@ func (s *Service) DetachChangesets(ctx context.Context, batchChangeID int64, ids
 
 	return nil
 }
-
-type ServiceMocks struct {
-	ValidateAuthenticator func(ctx context.Context, externalServiceID, externalServiceType string, a auth.Authenticator) error
-}
-
-func (sm ServiceMocks) Reset() {
-	sm.ValidateAuthenticator = nil
-}
-
-var Mocks ServiceMocks = ServiceMocks{}
